feat: Implement Joomla 5 module structure with dependency injection
- Restructured module to follow Joomla 5 namespaced architecture
- Implemented service provider and dependency injection container
- Migrated from assets/ to media/ folder with Web Asset Manager
- Added joomla.asset.json for proper asset registration
- Created namespaced Helper and Dispatcher classes
- Maintained backward compatibility with com_ajax bridge helper
- Fixed AJAX functionality with proper CSRF token handling
- Updated version to 1.0.1

This update modernizes the module to use Joomla 5's recommended patterns
while maintaining full functionality and improving maintainability.
