+++ services/provider.php
@@ -1,0 +1,35 @@
+<?php
+/**
+ * @package     Joomla.Site
+ * @subpackage  mod_whichpoliceforce
+ *
+ * @copyright   Copyright (C) 2025 SoftForge. All rights reserved.
+ * @license     GNU General Public License version 2 or later; see LICENSE
+ */
+
+defined('_JEXEC') or die;
+
+use Joomla\CMS\Extension\Service\Provider\HelperFactory;
+use Joomla\CMS\Extension\Service\Provider\Module;
+use Joomla\CMS\Extension\Service\Provider\ModuleDispatcherFactory;
+use Joomla\DI\Container;
+use Joomla\DI\ServiceProviderInterface;
+
+return new class () implements ServiceProviderInterface
+{
+    /**
+     * Registers the service provider with a DI container.
+     *
+     * @param   Container  $container  The DI container.
+     *
+     * @return  void
+     *
+     * @since   1.0.0
+     */
+    public function register(Container $container): void
+    {
+        $container->registerServiceProvider(new ModuleDispatcherFactory('\\Joomla\\Module\\WhichPoliceForce'));
+        $container->registerServiceProvider(new HelperFactory('\\Joomla\\Module\\WhichPoliceForce\\Site\\Helper'));
+        $container->registerServiceProvider(new Module());
+    }
+};
